Make ClientName unique per client; tie it to client id
Confusingly, HDFS has two ideas for keeping track of clients; the
client ID, which must be 16 bytes, and is passed with the RPC header;
and the client  "name", which can be longer, and is passed with some
miscellaneous RPC requests.

This change makes them tied to one another--the client name is a
(more) human-readable form of the ID. We also make sure to generate
a separate ID per client object, so hadoop doesn't get confused.