Fixed incorrect Append implementation.
The previous implementation considered existing blocks immutable, which is not
correct.

This commit attempts to unfinalize the last existing block during an append
operation so that it may be written to again. This is done by submitting the
correct stage `hdfs.OpWriteBlockProto_PIPELINE_SETUP_APPEND` during the
initial OpWriteBlock operation. Additionally, `blockWriteStream`s are now
always initialized with an offset, which will be > 0 in case of append
operation. Finally, the GenerationStamp field is incremented at the beginning
of the append operation, which is an incrementing integer that records the
version of a particular block.