Reorganize readdir code, fixing some performance problems
This reorganizes FileReader.Readdir and Client.ReadDir, changing the
latter to use the former and removing some redundant code. It also
changes the way the former works to be more performant.

Our strategy of batching (both in FileReader.Readdir and in the
complete command in the commandline client) was nonsensical, because
HDFS doesn't have any mechanism for requesting fewer entries at once.
To fetch a smaller "batch", we have to pull HDFS will give us (1000
at a time), and then throw out the rest. Worse, because we were
using 1024 as our batch size, we would fetch 2000 and throw out 974,
effectively making twice as many calls to HDFS every time.

Changing NeedLocation to false (which presumably means HDFS won't
attach block location to the entries) to false should also make these
calls faster.

Finally, a redundant test was removed and TestFileReadDirMany scaled
down so as to be a bit faster.