Add UseDatanodeHostname option, refactor BlockWriter and BlockReader
This adds the ability to specify whether the datanodes should be
connected to via hostname or IP address, rather than hardcoding
IP address.

To achieve this, the way BlockReaders and BlockWriters were created
had to be refactored; the New* methods are now deprecated, and the
structs are created manually. I think this is much cleaner, and
will make future changes easier.

One additional change was to move block finalization into FileWriter
and out of BlockWriter (since it involves communication with the
namenode, that was an odd place for it).