When seeking forward small amounts, don't reconnect
Currently, when seeking, we always discard the BlockReader and reconnect at the
new offset. This is often wasteful, but especially when seeking forward would
keep us in the same chunk, since on reconnecting we discard the beginning of the
chunk anyway.

This adds a new method to BlockReader, Skip, which attempts to discard bytes in
order to skip forward, and makes use of that for small (<64k) forward seeks.